ubdate payment and driver statistics

+++ app/src/main/java/gr/aueb/carpooling/model/view/passenger/ExistedSubroutes/ExistedSubrouteActivity.java
@@ -78,43 +78,60 @@ public class ExistedSubrouteActivity extends AppCompatActivity implements Existe
     @Override
 
     public void selectSubroute(Subroute subroute, Request_status status,Boolean b) {
-        showErrorMessage("Dest",subroute.getDestination().toString());
-        Money cost=subroute.calculateCost();
-        showErrorMessage("cost",String.valueOf(cost.getAmount()));
-        Boolean success= passenger.transaction(cost);
-        if(!success){
-            new androidx.appcompat.app.AlertDialog.Builder(this)
-                    .setCancelable(true)
-                    .setTitle("Η πληρωμή επέτυχε.Πρόσθεσε χρήματα.Η διαδρομη κόστησε")
-                    .setMessage(new DecimalFormat("0.00").format(cost.getAmount()))
-                    .setPositiveButton("OK", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            Intent intent = new Intent(ExistedSubrouteActivity.this, TopUpActivity.class);
-                            intent.putExtra("Username", username);
-                            startActivity(intent);
-                        }
-                    }).create().show();
-        }
+        //showErrorMessage("Dest",subroute.getDestination().toString());
+//        Money cost=subroute.calculateCost();
+//        showErrorMessage("cost",String.valueOf(cost.getAmount()));
+//        Boolean success= passenger.transaction(cost);
+//        if(!success){
+//            new androidx.appcompat.app.AlertDialog.Builder(this)
+//                    .setCancelable(true)
+//                    .setTitle("Η πληρωμή επέτυχε.Πρόσθεσε χρήματα.Η διαδρομη κόστησε")
+//                    .setMessage(new DecimalFormat("0.00").format(cost.getAmount()))
+//                    .setPositiveButton("OK", new DialogInterface.OnClickListener() {
+//                        @Override
+//                        public void onClick(DialogInterface dialog, int which) {
+//                            Intent intent = new Intent(ExistedSubrouteActivity.this, TopUpActivity.class);
+//                            intent.putExtra("Username", username);
+//                            startActivity(intent);
+//                        }
+//                    }).create().show();
+//        }
 
         if(b){
             if (status == Request_status.APPROVED){
-                Intent intent = new Intent(ExistedSubrouteActivity.this, DriverRatingActivity.class);
-                intent.putExtra("Username", username);
-                intent.putExtra("SubrouteId", subroute.getId());
-                startActivity(intent);
+                boolean success=payment(subroute);
+                if(success){
+                    subroute.setStatus(Request_status.COMPLETED);
+                    Intent intent = new Intent(ExistedSubrouteActivity.this, DriverRatingActivity.class);
+                    intent.putExtra("Username", username);
+                    intent.putExtra("SubrouteId", subroute.getId());
+                    startActivity(intent);
+                }else{
+                    Intent intent = new Intent(ExistedSubrouteActivity.this,TopUpActivity.class);
+                    intent.putExtra("Username", username);
+                    startActivity(intent);
+                }
+
             } else if (status == Request_status.REJECTED) {
                 showErrorMessage("You can press button complete if request status is approved. Current status ", String.valueOf(Request_status.REJECTED));
-            } else {
+            } else if(status == Request_status.PENDING){
                 showErrorMessage("Wait for driver's answer. Current status: ", String.valueOf(Request_status.PENDING));
+            } else if(status == Request_status.COMPLETED) {
+                showErrorMessage("You have already completed this subroute. Current status: ", String.valueOf(Request_status.COMPLETED));
             }
         } else {
-            Route route = routeDAO.findRouteBySubroute(subroute);
-            route.removeSubroute(subroute);
-            subrouteDAO.delete(subroute);
-            Intent intent = new Intent(this, this.getClass());
-            intent.putExtra("Username", username);
-            startActivity(intent);
+            if (status == Request_status.APPROVED) {
+                showErrorMessage("You cann't press button delete if request status is approved.Now is: ", String.valueOf(Request_status.APPROVED));
+            }else{
+                    Route route = routeDAO.findRouteBySubroute(subroute);
+                    route.removeSubroute(subroute);
+                    subrouteDAO.delete(subroute);
+                    Intent intent = new Intent(this, this.getClass());
+                    intent.putExtra("Username", username);
+                    startActivity(intent);
+            }
+
+
         }
 
     }
@@ -152,6 +169,19 @@ public class ExistedSubrouteActivity extends AppCompatActivity implements Existe
 ////        intent.putExtra("Username",username);
 //        startActivity(intent);
 
+
+    }
+
+    @Override
+    public boolean payment(Subroute subroute) {
+        showErrorMessage("Balance",new DecimalFormat("0.00").format(passenger.getBalance()));
+        Money cost=subroute.calculateCost();
+        boolean success= passenger.transaction(cost);
+        if(!success){
+            showErrorMessage("Η πληρωμή επέτυχε.Πρόσθεσε χρήματα.Η διαδρομη κόστησε",new DecimalFormat("0.00").format(cost.getAmount()));
+        }
+
+        return success;
 
     }
